[1-5] refactor: 패키지 정리
