+++ src/main/java/carRacing/RacingApplication.java
@@ -1,15 +1,9 @@
 package carRacing;
 
+import carRacing.controller.CarRacingController;
+
 public class RacingApplication {
     public static void main(String[] args) {
-        // 유저로부터 파라미터 입력
-        InputView inputView = new InputView();
-
-        // 게임 초기화 및 경주 진행
-        CarRacingGame carRacingGame = new CarRacingGame(inputView.getCarNamesString(), inputView.getTryCount());
-        RacingHistory racingHistory = carRacingGame.doGame();
-
-        // 결과 출력
-        ResultView.displayGameResult(racingHistory);
+        new CarRacingController();
     }
 }
