+++ src/test/java/carRacing/CarTest.java
@@ -1,5 +1,6 @@
 package carRacing;
 
+import carRacing.domain.Car;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
