+++ src/test/java/carRacing/RacingHistoryTest.java
@@ -1,5 +1,7 @@
 package carRacing;
 
+import carRacing.domain.Car;
+import carRacing.domain.RacingHistory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
